logical struct
