+++ polars/polars-core/src/series/implementations/boolean.rs
@@ -32,8 +32,11 @@ impl IntoSeries for BooleanChunked {
 }
 
 impl private::PrivateSeries for SeriesWrap<BooleanChunked> {
-    fn _field(&self) -> &Field {
-        self.0.ref_field()
+    fn _field(&self) -> Cow<Field> {
+        Cow::Borrowed(self.0.ref_field())
+    }
+    fn _dtype(&self) -> &DataType {
+        self.0.ref_field().data_type()
     }
 
     fn explode_by_offsets(&self, offsets: &[i64]) -> Series {
@@ -42,7 +45,7 @@ impl private::PrivateSeries for SeriesWrap<BooleanChunked> {
 
     #[cfg(feature = "asof_join")]
     fn join_asof(&self, other: &Series) -> Result<Vec<Option<u32>>> {
-        self.0.join_asof(other.as_ref().as_ref())
+        self.0.join_asof(other)
     }
 
     fn set_sorted(&mut self, reverse: bool) {
