+++ polars/polars-core/src/series/implementations/object.rs
@@ -31,10 +31,9 @@ impl<T> PrivateSeries for SeriesWrap<ObjectChunked<T>>
 where
     T: PolarsObject,
 {
-    fn _field(&self) -> &Field {
-        self.0.ref_field()
+    fn _field(&self) -> Cow<Field> {
+        Cow::Borrowed(self.0.ref_field())
     }
-
     fn agg_first(&self, _groups: &[(u32, Vec<u32>)]) -> Series {
         self.0.agg_first(_groups)
     }
