+++ polars/polars-core/src/series/into.rs
@@ -1,0 +1,31 @@
+use crate::prelude::*;
+use arrow::compute::cast::cast;
+
+impl Series {
+    /// Convert a chunk in the Series to the correct Arrow type.
+    /// This conversion is needed because polars doesn't use a
+    /// 1 on 1 mapping for logical/ categoricals, etc.
+    pub(crate) fn to_arrow(&self, chunk_idx: usize) -> ArrayRef {
+        match self.dtype() {
+            #[cfg(feature = "dtype-categorical")]
+            DataType::Categorical => {
+                let ca = self.categorical().unwrap();
+                let ca = CategoricalChunked::new_from_chunks(
+                    ca.name(),
+                    vec![ca.chunks()[chunk_idx].clone()],
+                );
+                let arr: DictionaryArray<u32> = (&ca).into();
+                Arc::new(arr) as ArrayRef
+            }
+            DataType::Date32 => {
+                let arr = cast(&*self.chunks()[chunk_idx], &DataType::Date32.to_arrow()).unwrap();
+                Arc::from(arr)
+            }
+            DataType::Date64 => {
+                let arr = cast(&*self.chunks()[chunk_idx], &DataType::Date64.to_arrow()).unwrap();
+                Arc::from(arr)
+            }
+            _ => self.chunks()[chunk_idx].clone(),
+        }
+    }
+}
